fix toggleable variable in config to set the docked mode

+++ Ryujinx.HLE/OsHle/Services/Am/ICommonStateGetter.cs
@@ -20,14 +20,14 @@ namespace Ryujinx.HLE.OsHle.Services.Am
         {
             m_Commands = new Dictionary<int, ServiceProcessRequest>()
             {
-                { 0, GetEventHandle                          },
-                { 1, ReceiveMessage                          },
-                { 5, GetOperationMode                        },
-                { 6, GetPerformanceMode                      },
-                { 8, GetBootMode                             },
-                { 9, GetCurrentFocusState                    },
-                { 60, GetDefaultDisplayResolution            },
-                { 61, GetDefaultDisplayResolutionChangeEvent }
+                { 0,  GetEventHandle                          },
+                { 1,  ReceiveMessage                          },
+                { 5,  GetOperationMode                        },
+                { 6,  GetPerformanceMode                      },
+                { 8,  GetBootMode                             },
+                { 9,  GetCurrentFocusState                    },
+                { 60, GetDefaultDisplayResolution             },
+                { 61, GetDefaultDisplayResolutionChangeEvent  }
             };
 
             DisplayResolutionChangeEvent = new KEvent();
@@ -69,7 +69,7 @@ namespace Ryujinx.HLE.OsHle.Services.Am
         {
             Apm.PerformanceMode Mode = DockedMode ? Apm.PerformanceMode.Docked : Apm.PerformanceMode.Handheld;
 
-            Context.ResponseData.Write((byte)Mode);
+            Context.ResponseData.Write((int)Mode);
 
             return 0;
         }
